packages/threat-composer-ai/pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".

[project]
name = "threat-composer-ai"
version = "0.1.1"
description = "AI-powered automated threat modeling for codebases"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "boto3",
  "botocore",
  "click",
  "fastmcp",
  "json-schema-to-pydantic",
  "opentelemetry-api",
  "opentelemetry-exporter-otlp-proto-http",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-propagator-aws-xray",
  "opentelemetry-sdk",
  "pathspec",
  "psutil",
  "rich",
  "strands-agents",
  "strands-agents-tools"
]

  [project.scripts]
  threat-composer-ai-cli = "threat_composer_ai.__main__:main"
  threat-composer-ai-mcp = "threat_composer_ai.mcp_server:main"

[dependency-groups]
dev = [
  "pytest==7.4.3",
  "pytest-mock",
  "pytest-cov",
  "pytest-asyncio",
  "ruff"
]

[tool.uv]
managed = true
package = true

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
  ".git",
  "__pycache__",
  ".venv",
  ".eggs",
  "*.egg",
  "dist",
  "build"
]

  [tool.ruff.lint]
  select = [ "E", "W", "F", "I", "B", "C4", "UP" ]
  ignore = [ "E501" ]

  [tool.ruff.format]
  quote-style = "double"
  indent-style = "space"
  skip-magic-trailing-comma = false
  line-ending = "auto"
